requirements.txt
cycler==0.10.0
Cython==0.29.7
kiwisolver==1.0.1
matplotlib==3.0.3
moog==0.1
numpy==1.16.3
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
periodictable==1.5.0
